fix: fixed manual calendar addition
