+++ src/addCellTimetable.js
@@ -1,5 +1,7 @@
+const { create } = require("domain");
+
 // function calls method to create a cell object then adds that object to the inputted array obj
-function addCellTimetable(id, user, userCustomObj) {
+function addCellTimetable(id, user) {
   const cell = createCellInstance(id, user, 1);
   userCustomObj.cells.push(cell); // adds cell to inputted array
   return userCustomObj;
@@ -7,10 +9,10 @@ function addCellTimetable(id, user, userCustomObj) {
 
 function createCellInstance(id, user, numPeople) {
   return {
-    ID: id,
+    id: id,
     users: [user],
     numPeople: numPeople,
   };
 }
 
-module.exports = addCellTimetable;
+module.exports = createCellInstance;
